rename app config file
